+++ collectors/bzimport/constants.py
@@ -27,3 +27,5 @@ ANALYSIS_TASK_PRODUCT = "Security Response"
 # Bugzilla datetime format strings
 BZ_DT_FMT = "%Y-%m-%dT%H:%M:%S%z"
 BZ_DT_FMT_HISTORY = "%Y-%m-%dT%H:%M:%SZ"
+
+FLAW_PLACEHOLDER_KEYWORD = "Tracking"