+++ docs/CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Implement validation for embargoed flaws not be able to have public trackers (OSIDB-350)
 - Fix Jira tracker created and updated timestamps (OSIDB-14)
 - Fix errata created and updated timestamps (OSIDB-453)
+- Restrict write operations on placeholder flaws (OSIDB-388)
 
 ## [2.3.4] - 2022-12-15
 ### Changed